// src/app/enums/event-type.enum.ts
export enum EventType {
  LoginAttempt = 'LOGIN_ATTEMPT',
  LoginAttemptFailure = 'LOGIN_ATTEMPT_FAILURE',
  LoginAttemptSuccess = 'LOGIN_ATTEMPT_SUCCESS',
  ProfileUpdate = 'PROFILE_UPDATE',
  ProfilePictureUpdate = 'PROFILE_PICTURE_UPDATE',
  RoleUpdate = 'ROLE_UPDATE',
  AccountSettingsUpdate = 'ACCOUNT_SETTINGS_UPDATE',
  PasswordUpdate = 'PASSWORD_UPDATE',
  MfaUpdate = 'MFA_UPDATE'
}